requirements.txt
asgiref==3.2.10
Django==3.0.8
django-timezone-field==4.0
djangorestframework==3.11.0
gunicorn==20.0.4
pytz==2020.1
sqlparse==0.3.1
